Authentication middleware

+++ app/Http/Controllers/StudentController.php
@@ -4,30 +4,39 @@ namespace App\Http\Controllers;
 
 use App\Models\Majors;
 use App\Models\Student;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class StudentController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-student');
+
         $students = Student::with('majors')->get();
         return view('students.index', compact('students'));
     }
 
     public function show(string $id)
     {
+        Gate::authorize('view-student');
+
         $student = Student::with('majors')->find($id);
         return view('students.show', compact('student'));
     }
 
     public function create()
     {
+        Gate::authorize('store-student');   
+
         $majors = Majors::all();
         return view('students.create', compact('majors'));
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('store-student');
+
         $validated = $request->validate([
             'name' => 'required',
             'student_id_number' => 'required|unique:students|max:9',
@@ -52,8 +61,11 @@ class StudentController extends Controller
 
         return redirect()->route('students.index')->with('success', 'Student created successfully');
     }
+    
     public function edit(string $id)
         {
+            Gate::authorize('edit-student');
+
             $student = Student::with('majors')->find($id);
             $majors = Majors::all();
             return view('students.edit', compact('student', 'majors'));
@@ -61,6 +73,8 @@ class StudentController extends Controller
 
         public function update(Request $request, string $id)
         {
+            Gate::authorize('edit-student');
+
             $validated = $request->validate([
                 'name' => 'required',
                 'student_id_number' => "required|unique:students,student_id_number,$id|max:9",
@@ -90,6 +104,8 @@ class StudentController extends Controller
 
         public function destroy(string $id)
         {
+            Gate::authorize('destroy-student');
+
             $student = Student::findOrFail($id);
 
             $student->delete();
